requirements.txt
# 업비트 API 연동 및 JWT 생성
requests
pyjwt

# 데이터 분석 및 지표 계산
pandas
ta

# AI (LLM) 연동
openai
google-generativeai

# 환경 변수 관리를 위한 라이브러리
python-dotenv

# Web framework for runtime API
fastapi
uvicorn

# Streamlit UI
streamlit
